Fix fast refresh compatibility for discussion surveys
