+++ frontend/src/employee-frontend/components/unit/tab-calendar/discussion-surveys/times-calendar/TimesCalendar.tsx
@@ -85,7 +85,7 @@ function groupByWeek(days: DiscussionReservationDay[]): CalendarWeek[] {
   return weeks
 }
 
-export const groupByMonth = (
+const groupByMonth = (
   dailyData: DiscussionReservationDay[]
 ): CalendarMonth[] => {
   const getWeekMonths = (weeklyData: CalendarWeek) => {
